Add files via upload
I'm just a beginner in the world of WEB DEVELOPMENT and I Started my Web Development journey by beginning with learning HTML and CSS that are the fundamentals for frontend development. This is a mini portfolio that I've created using the basics of HTML and CSS.
Looking forward to create more projects using the latest technologies.